requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
python-dotenv==1.0.0
pandas>=2.1.3
numpy>=1.26.2
pandas_ta>=0.4.0   # updated to a version compatible with Python 3.11
alpaca-trade-api>=5.0.0
celery>=5.3.0
redis>=5.3.0
sqlalchemy>=2.1.0
python-multipart>=0.0.6
bcrypt>=4.1.0
pyjwt>=2.8.0
alembic>=1.13.0